+++ client/web/src/organisms/mainpage/ranking/ToptenCard.tsx
@@ -8,7 +8,7 @@ import SentimentSatisfiedAltIcon from '@material-ui/icons/SentimentSatisfiedAlt'
 import SentimentDissatisfiedIcon from '@material-ui/icons/SentimentDissatisfied';
 import useAxios from 'axios-hooks';
 import { useTopTenCard, useTabs, useTabItem } from './style/TopTenCard.style';
-import ToptenList from './topten/TopTenList';
+import TopTenList from './topten/TopTenList';
 
 const columns = [
   { name: 'admire', label: '감탄점수', icon: <SentimentVerySatisfiedIcon /> },
@@ -61,7 +61,11 @@ function TopTenCard(): JSX.Element {
           </Tabs>
         </Grid>
         <Grid item xs={9}>
-          <ToptenList data={data} currentTab={columns[tabIndex].name} loading={loading} />
+          <TopTenList
+            data={data}
+            currentTab={columns[tabIndex].name}
+            loading={loading}
+          />
         </Grid>
       </Grid>
     </section>
